requirements.txt
google-generativeai>=0.5.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyautogui>=0.9.54
psutil>=5.9.0
requests>=2.31.0
schedule>=1.2.0
websocket-client>=1.7.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
obs-websocket-py>=1.0.0
pygetwindow>=0.0.9
